Upgrade tooling (#815)
* Bump rustc to 1.68 and appease clippy.

* Bump stackage snapshot to lts-20.16.

* Remove dead code.

+++ concordium-node/src/connection/low_level.rs
@@ -263,7 +263,7 @@ impl ConnectionLowLevel {
     fn initialize(&mut self) {
         // Set linger time if requested
         if let Some(linger) = self.so_linger {
-            self.set_linger(true, linger as u16);
+            self.set_linger(true, linger);
         }
 
         if let Err(e) = self.socket.set_nodelay(true) {
